+++ extension/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "extension",
+  "name": "virtual-json-viewer",
   "version": "0.0.2",
   "private": true,
   "dependencies": {
