+++ extension/src/viewer/components/Toolbar/SearchBox.tsx
@@ -1,7 +1,7 @@
 import classNames from "classnames";
 import { Dispatch, FormEvent, SetStateAction, useEffect, useRef } from "react";
 import { Icon, IconButton, IconLabel } from "viewer/components";
-import { Search } from "../../commons/state";
+import { Search } from "../../state";
 
 // TODO setting typing delay
 const TYPING_DELAY = 300;
