+++ extension/src/viewer/components/TreeViewer/TreeViewer.tsx
@@ -5,8 +5,8 @@ import {
   VariableSizeTree as Tree,
 } from "react-vtree";
 import { EventType } from "viewer/commons/EventBus";
-import { Search } from "viewer/commons/state";
 import { useEventBusListener, useWindowSize } from "viewer/hooks";
+import { Search } from "viewer/state";
 import { JsonNodeData } from "./model/JsonNode";
 import { getRootNodes, isLeaf, jsonTreeWalker } from "./model/JsonTreeWalker";
 import { TreeNode } from "./TreeNode";
