+++ extension/src/viewer/components/TreeViewer/model/JsonTreeWalker.ts
@@ -1,6 +1,6 @@
 import { TreeWalker, TreeWalkerValue } from "react-vtree";
 import * as J from "viewer/commons/JsonUtils";
-import { Search } from "../../../commons/state";
+import { Search } from "../../../state";
 import { JsonNode, JsonNodeData, SearchMatch } from "./JsonNode";
 import { NodeFilter } from "./NodeFilter";
