+++ extension/src/viewer/hooks/UseStorage.ts
@@ -1,5 +1,5 @@
 import { Dispatch, useContext, useEffect, useMemo, useState } from "react";
-import { Runtime, RuntimeContext } from "viewer/commons/state";
+import { Runtime, RuntimeContext } from "viewer/state";
 import { Mutex, useEffectAsync } from ".";
 
 export function useStorage<T>(key: string, defaultValue: T): [T, Dispatch<T>] {
