+++ extension/src/viewer/state/Search.ts
