+++ extension/src/viewer/state/ViewerMode.ts
